// my_investmentplan_frontend/src/Types/AdminDTO/UnverifiedInvestor.ts
export interface InvestorDetail {
  invId: number;
  publicInvestorId: string;
  investorImage: string; 
  mobile: string;
  aadhaarNo: string;
  panNo: string;
  age: number;
  signature: string; 
  incomeProof: string; 
  bankName: string;
  fund: number;
  signedDocument: string; 
  nomineeName: string;
  nomineeEmail: string;
  nomineeRelation: string;
}

export interface UnverifiedInvestor {
  publicInvestorId: string;
  investorName: string;
  email: string;
  password: string;
  verificationStatus: number;
  status: number;
  roleId: number;
  createdAt: string;
  investorDetail: InvestorDetail;
}

export type UnverifiedInvestorListResponse = UnverifiedInvestor[];